# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plume",
    "version": "0.1.0",
    "description": "API REST de Plume. Toutes les requêtes et réponses sont en JSON (sauf mention). Les requêtes qui modifient un état exigent la session et l’en-tête X-CSRF-Token (jeton renvoyé par /auth/login et /auth/session), ainsi qu’un en-tête Origin identique à l’URL publique."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "postAuthLogin",
        "summary": "Connexion (identifiants IMAP)",
        "tags": [
          "auth"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 320
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024
                  },
                  "remember": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/totp/verify": {
      "post": {
        "operationId": "postAuthTotpVerify",
        "summary": "Second facteur TOTP ou code de secours",
        "tags": [
          "auth"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string",
                        "maxLength": 16
                      }
                    },
                    "required": [
                      "code"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "backupCode": {
                        "type": "string",
                        "maxLength": 32
                      }
                    },
                    "required": [
                      "backupCode"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/auth/session": {
      "get": {
        "operationId": "getAuthSession",
        "summary": "État de la session",
        "tags": [
          "auth"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "postAuthLogout",
        "summary": "Déconnexion",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "204": {
            "description": "Succès, sans contenu"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/auth/logout-all": {
      "post": {
        "operationId": "postAuthLogoutAll",
        "summary": "Déconnexion de toutes les sessions",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "204": {
            "description": "Succès, sans contenu"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/auth/accounts": {
      "post": {
        "operationId": "postAuthAccounts",
        "summary": "Ajoute un compte à la session (mêmes contrôles qu’une connexion)",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 320
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024
                  },
                  "remember": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/accounts/totp": {
      "post": {
        "operationId": "postAuthAccountsTotp",
        "summary": "Second facteur du compte en cours d’ajout",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string",
                        "maxLength": 16
                      }
                    },
                    "required": [
                      "code"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "backupCode": {
                        "type": "string",
                        "maxLength": 32
                      }
                    },
                    "required": [
                      "backupCode"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/auth/accounts/cancel": {
      "post": {
        "operationId": "postAuthAccountsCancel",
        "summary": "Abandonne l’ajout d’un compte",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "204": {
            "description": "Succès, sans contenu"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/auth/accounts/remove": {
      "post": {
        "operationId": "postAuthAccountsRemove",
        "summary": "Ferme un compte de la session",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 320
                  }
                },
                "required": [
                  "email"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/switch": {
      "post": {
        "operationId": "postAuthSwitch",
        "summary": "Bascule vers un autre compte de la session",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 320
                  }
                },
                "required": [
                  "email"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "operationId": "getMe",
        "summary": "Utilisateur courant",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/me/totp/setup": {
      "post": {
        "operationId": "postMeTotpSetup",
        "summary": "Prépare l’activation TOTP",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/me/totp/enable": {
      "post": {
        "operationId": "postMeTotpEnable",
        "summary": "Active TOTP",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "maxLength": 16
                  }
                },
                "required": [
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/me/totp/disable": {
      "post": {
        "operationId": "postMeTotpDisable",
        "summary": "Désactive TOTP",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "204": {
            "description": "Succès, sans contenu"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "maxLength": 16
                  }
                },
                "required": [
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/me/totp/backup-codes": {
      "post": {
        "operationId": "postMeTotpBackupCodes",
        "summary": "Régénère les codes de secours",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "maxLength": 16
                  }
                },
                "required": [
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/folders": {
      "get": {
        "operationId": "getFolders",
        "summary": "Dossiers et compteurs",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/labels": {
      "get": {
        "operationId": "getLabels",
        "summary": "Libellés",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/messages/{id}/draft": {
      "get": {
        "operationId": "getMessagesIdDraft",
        "summary": "Corps d’un brouillon, pour le reprendre",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1500,
              "description": "Identifiant opaque de message"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-plume_sid"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "responses": {
      "Error": {
        "description": "Erreur (code stable, sans détail interne)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              },
              "required": [
                "error"
              ]
            }
          }
        }
      }
    }
  }
}